+++ html5-graphics/src/main/scala/com/funlabyrinthe/graphics/html/Animated.scala
@@ -5,8 +5,8 @@ import com.funlabyrinthe.core.graphics.Image
 final class Animated(val frames: IArray[Image]) extends Image:
   def isComplete: Boolean = true
 
-  val width: Double = frames.head.width
-  val height: Double = frames.head.height
+  val width: Int = frames.head.width
+  val height: Int = frames.head.height
 
   def isAnimated: Boolean = true
   val time = frames.map(_.time).sum
